move code from lazy-container to lazy-table-block
